+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.0.0
 =======
 
+* Added an example app to demonstrate Local Video Snapshot.
 * Added an example app to demonstrate Media Device Selection.
 * Updated twilio-video to 1.0.0
 * Removed configuration profile